Revert changes to test fixtures
