+++ tests/test_lexer.py
@@ -4,6 +4,12 @@ import lkml
 import lkml.tokens as tokens
 
 
+@pytest.fixture
+def lexer():
+    text = "Some sample text for testing."
+    return lkml.Lexer(text)
+
+
 def test_peek_does_not_advance_index(lexer):
     index = lexer.index
     lexer.peek()
